+++ src/currents/current.rs
@@ -9,7 +9,7 @@ use derive_more::Display;
 use enum_dispatch::enum_dispatch;
 use nalgebra::{Point3, Vector3};
 
-use super::{CircularCurrent, CurrentPath};
+use super::{CircularCurrent, PathCurrent};
 use crate::base::{Float, Pose, Source, Transform};
 
 /// Current source variants.
@@ -18,5 +18,5 @@ use crate::base::{Float, Pose, Source, Transform};
 #[enum_dispatch(Source<T>, Transform<T>,)]
 pub enum Current<T: Float = f64> {
     Circular(CircularCurrent<T>),
-    Path(CurrentPath<T>),
+    Path(PathCurrent<T>),
 }