+++ src/fields/mod.rs
@@ -67,18 +67,18 @@
 
 mod field_circular;
 mod field_cuboid;
-mod field_current_path;
 mod field_cylinder;
 mod field_dipole;
+mod field_path_current;
 mod field_sphere;
 mod field_tetrahedron;
 mod field_triangle;
 
 pub use field_circular::{circular_B, circular_B_batch, sum_multiple_circular_B};
 pub use field_cuboid::{cuboid_B, cuboid_B_batch, sum_multiple_cuboid_B};
-pub use field_current_path::{current_path_B, current_path_B_batch, sum_multiple_currentpath_B};
 pub use field_cylinder::{cylinder_B, cylinder_B_batch, sum_multiple_cylinder_B};
 pub use field_dipole::{dipole_B, dipole_B_batch, sum_multiple_dipole_B};
+pub use field_path_current::{path_current_B, path_current_B_batch, sum_multiple_path_current_B};
 pub use field_sphere::{sphere_B, sphere_B_batch, sum_multiple_sphere_B};
 pub use field_tetrahedron::{sum_multiple_tetrahedron_B, tetrahedron_B, tetrahedron_B_batch};
 pub use field_triangle::{sum_multiple_triangle_B, triangle_B, triangle_B_batch};
@@ -97,7 +97,7 @@ crate::crate_utils::need_unstable! {
     pub use field_sphere::local_sphere_B;
     pub use field_tetrahedron::{local_tetrahedron_B, tetrahedron_B_precomputed, precompute_tetrahedron};
     pub use field_triangle::local_triangle_B;
-    pub use field_current_path::local_current_path_B;
+    pub use field_path_current::local_path_current_B;
 
     #[cfg(feature = "mesh")]
     pub use field_mesh::local_mesh_B;